perf(database): Improve text search lookup among other features
1) The production server has some problems when looking up signatures by
their text and kind, i.e. `balanceOf`. The reason for this is because we join the
`signature` table with `mapping_signature_github`. As of now
`mapping_signature_github` has ~52 million entries and as such yields really
bad performance. Furthermore joining this with `mapping_signature_fourbyte` or
`mapping_signature_etherscan` is impossible because timings are in the minutes range.
To fix this we introduce a `mapping_signature_kind` table which simply acts as a
n-m table for `signature` to find signatures by their kind and id / text.
Not only does it improve the performance (because easily indexable) but also
allows us to find signatures regardless of their source, i.e. GitHub, Etherscan
or Fourbyte.

2) The signature table now contains a `is_valid` field which is used to
mark signatures with custom types, i.e. `balanceOf(ERC20)`. The features
for this in the backend is not implemented yet however and will be
introduces in a later commit.